fix subtle possible race that could kill another process and simplify code
Fri Aug 23 01:09:36 2002  George Lebl <jirka@5z.com>

	* gui/gdmlogin.c: fix subtle possible race that could kill another
	  process and simplify code in the meantime.  Ignore SIGCHLD, and
	  don't wait for the background process until we want to kill it,
	  that may leave a zombie for a little bit if the process dies early,
	  but who cares.

	* daemon/display.c, daemon/gdm.c, daemon/misc.c, daemon/server.c,
	  daemon/slave.c: Fix many many very minor races but such that
	  could conceivably take things down (a kill (0,...) "could"
	  very theoretically occur previously).  Fix some worse but still
	  not very likely races by using the push/pop of sigchld block in
	  places where we blocked by hand.  Ensure all processes are always
	  killed on slave exit (probably doing things redundantly in some
	  cases).  Fix some possible cases of syslog from a signal which
	  does bad things to our karma sometimes.  When the server dies
	  make sure we don't do anything with the display (we null the d->dsp
	  var) to make sure we don't reenter the xioerror_handler.  The
	  extra_process uses 0 and not -1 as the no value everywhere.
	  Make sure kills are never called with -1 (probably just anal,
	  but one never knows).  Do setsid and the init in daemonize
	  on -nodaemon as well.  Do setsid for greeter/chooser/config.

+++ daemon/server.c
@@ -113,7 +113,10 @@ gdm_server_reinit (GdmDisplay *disp)
 
 	gdm_debug ("gdm_server_reinit: Server for %s is about to be reinitialized!", disp->name);
 
-	kill (disp->servpid, SIGHUP);
+	gdm_sigchld_block_push ();
+	if (disp->servpid > 0)
+		kill (disp->servpid, SIGHUP);
+	gdm_sigchld_block_pop ();
 
 	/* HACK! the Xserver can't really tell us when it got the hup signal,
 	 * so we are really stuck just going to sleep for a bit hoping that
@@ -150,7 +153,7 @@ gdm_server_stop (GdmDisplay *disp)
 
     disp->servstat = SERVER_DEAD;
 
-    if (disp->servpid != 0) {
+    if (disp->servpid > 0) {
 	    pid_t servpid;
 
 	    gdm_debug ("gdm_server_stop: Killing server pid %d",
@@ -536,6 +539,8 @@ gdm_server_start (GdmDisplay *disp, gboolean treat_as_flexi,
     /* bad things are happening */
     if (d->servpid > 0) {
 	    pid_t pid;
+
+	    d->dsp = NULL;
 
 	    gdm_sigchld_block_push ();
 	    pid = d->servpid;
@@ -1048,7 +1053,7 @@ gdm_server_child_handler (int signal)
 			write (server_signal_pipe[1], "Yay!", 4);
 		} else if (pid == extra_process) {
 			/* an extra process died, yay! */
-			extra_process = -1;
+			extra_process = 0;
 			extra_status = status;
 		}
 	}
